server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.21",
    "langgraph>=0.3.29",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
